Adds stats reporting

+++ front_end/generate_tree/index.php
@@ -70,7 +70,7 @@ if(!array_key_exists('payload', $_POST) || $_POST['payload'] == '')
 		$include_common_names,
 		$include_authors,
 		$include_sources,
-		$fill_in_sources,
+		$fill_in_links,
 		$use_file_splitter
 	]
 ] = json_decode($_POST['payload'], TRUE);
@@ -146,12 +146,45 @@ foreach($ranks[$kingdom] as $rank_id => $rank_data){
 	if($include_common_names)
 		$line .= $column_separator . $rank_name . ' Common Name';
 
-	if($include_sources || $fill_in_sources)
+	if($include_sources || $fill_in_links)
 		$line .= $column_separator . $rank_name . ' Source';
 
 }
 
 $header_line = $line . $line_separator;
+
+
+//send stats data
+if(STATS_URL!=''){
+
+	$friendly_selected_ranks = [];
+	foreach($selected_ranks as $rank_id)
+		$friendly_selected_ranks[] = $ranks[$kingdom][$rank_id][0];
+
+	$stats_data = [
+		'site'    => 'gbif',
+		'ip'      => $_SERVER['REMOTE_ADDR'],
+		'tree'    => $choice_tree,
+		'ranks'   => $friendly_selected_ranks,
+		'options' => [
+			'include_common_names' => $include_common_names,
+			'include_authors'      => $include_authors,
+			'include_sources'      => $include_sources,
+			'fill_in_links'        => $fill_in_links,
+			'use_file_splitter'    => $use_file_splitter
+		],
+	];
+
+	$options = [
+		'http' => [
+			'header'  => "Content-type: application/x-www-form-urlencoded\r\n",
+			'method'  => 'POST',
+			'content' => http_build_query($stats_data)
+		]
+	];
+	$context = stream_context_create($options);
+	$result = file_get_contents(STATS_URL, FALSE, $context);
+}
 
 
 //Output the data
@@ -164,6 +197,7 @@ if($use_file_splitter)
 	$line_limit = 7000;
 else
 	$line_limit = FALSE;
+
 
 function show_node(
 	$taxon_number,
@@ -181,7 +215,7 @@ function show_node(
 	global $kingdom;
 	global $ranks;
 	global $selected_ranks;
-	global $fill_in_sources;
+	global $fill_in_links;
 	global $result;
 	global $line_limit;
 	global $lines_count;
@@ -217,7 +251,7 @@ function show_node(
 
 		if($include_sources && $node[0][3]!='')
 			$line .= $column_separator.$node[0][3];
-		elseif($fill_in_sources && $node[0][3]=='')
+		elseif($fill_in_links && $node[0][3]=='')
 			$line .= $column_separator.'https://gbif.gov/species/'.$taxon_number;
 
 		$result .= $line . $line_separator;
@@ -248,7 +282,7 @@ function handle_missing_ranks(
 	global $include_common_names;
 	global $include_sources;
 	global $required_ranks;
-	global $fill_in_sources;
+	global $fill_in_links;
 
 	if($rank==0)
 		return '';
@@ -267,7 +301,7 @@ function handle_missing_ranks(
 		if($include_common_names)
 			$count++;
 
-		if($include_sources || $fill_in_sources)
+		if($include_sources || $fill_in_links)
 			$count++;
 
 		$line .= str_repeat($column_separator, $count);
@@ -372,9 +406,11 @@ else {
 	}
 
 
-	foreach (glob($target_dir.'*.*') as $file_name)
-		unlink($file_name);
+	if($target_dir!==''){
+		foreach (glob($target_dir.'*.*') as $file_name)
+			unlink($file_name);
 
-	rmdir($target_dir);
+		rmdir($target_dir);
+	}
 
 }
